allow aws credential-providers to be tree-shaken
